Move setting of logged-in users's auth token the the IAuthApi companion object
