+++ app/src/main/java/com/realityexpander/tasky/data/repository/remote/IAuthApi.kt
@@ -1,5 +1,6 @@
 package com.realityexpander.tasky.data.repository.remote
 
+import com.realityexpander.tasky.BuildConfig
 import com.realityexpander.tasky.common.AuthToken
 import com.realityexpander.tasky.common.Email
 import com.realityexpander.tasky.common.Password
@@ -19,4 +20,22 @@ interface IAuthApi {
     )
 
     suspend fun authenticate(authToken: AuthToken? = null): Boolean
+
+    companion object {
+        var authToken: String? = null
+            private set
+
+        var authorizationHeader: String? = null
+            private set
+            get() = "Bearer ${Companion.authToken}"
+
+        @JvmName("setAuthToken1")
+        fun setAuthToken(authToken: AuthToken?) {
+            Companion.authToken = authToken
+        }
+
+        fun createAuthorizationHeader(authToken: String?): String {
+            return "Bearer $authToken"
+        }
+    }
 }
